refactor: move renovate init config helper Signed-off-by: Gregor Zeitlinger <gregor.zeitlinger@grafana.com>
